Minor Fixes
- Enabled bloom clamp to avoid a HDR error in the sun on some platforms.
- Changed directives in XRGraphics (in Core RP) to better support platforms that don't use XR at all.
- Fixed a float4/half4 issue in the Universal RP DOF shader.

+++ Packages/com.unity.render-pipelines.universal/Shaders/PostProcessing/BokehDepthOfField.shader
@@ -127,7 +127,7 @@ Shader "Hidden/Universal Render Pipeline/BokehDepthOfField"
             return half4(avg, coc);
         }
 
-        void Accumulate(float4 samp0, float2 uv, float2 disp, inout float4 farAcc, inout float4 nearAcc)
+        void Accumulate(float4 samp0, float2 uv, float2 disp, inout half4 farAcc, inout half4 nearAcc)
         {
             float dist = length(disp);
